pyproject.toml
[tool.poetry]
name = "ot-harjoitustyo"
version = "0.1.0"
description = ""
authors = ["Olli-Pekka Turtio <olli-pekka.turtio@helsinki.fi>"]

[tool.poetry.dependencies]
python = "^3.8"
cowsay = "^5.0"
invoke = "^1.7.3"
pytest = "^7.2.0"
python-dotenv = "^0.21.0"
pillow = "^9.3.0"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
coverage = "^6.4.2"
pytest-dotenv = "^0.5.2"

[tool.poetry.group.dev.dependencies]
autopep8 = "^2.0.0"
pillow = "^9.3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
